Config: Move JSON Import/Export to parameters
